Develop

+++ src/code/canvas/utilities/arrays/dot/dot.ts
@@ -3,14 +3,14 @@ import { ArrayBase } from "../base.js";
 export class DotArray extends ArrayBase {
     private _dotsX: Int16Array;
     private _dotsY: Int16Array;
-    private _radiuses: Float32Array;
+    private _radiuses: Int16Array;
     private _colors: Array<string>;
 
     constructor() {
         super();
         this._dotsX = new Int16Array(this.space);
         this._dotsY = new Int16Array(this.space);
-        this._radiuses = new Float32Array(this.space);
+        this._radiuses = new Int16Array(this.space);
         this._colors = new Array<string>();
     }
 
@@ -22,7 +22,7 @@ export class DotArray extends ArrayBase {
         return this._dotsY.slice(0, this.length);
     }
 
-    public get radiuses(): Readonly<Float32Array> {
+    public get radiuses(): Readonly<Int16Array> {
         return this._radiuses.slice(0, this.length);
     }
 
@@ -65,7 +65,7 @@ export class DotArray extends ArrayBase {
 
     private expandRadiuses(): void {
         const radiuses = this._radiuses;
-        this._radiuses = new Float32Array(this.space);
+        this._radiuses = new Int16Array(this.space);
 
         for (let index = 0; index < radiuses.length; index++) {
             this._radiuses[index] = radiuses[index];
